refactor(utils): change utils file names to kebab case, remove/merge utils files
