+++ src/components/InfoDisplay/TagsDisplay.tsx
@@ -1,6 +1,6 @@
 import { Badge, BadgeVariants } from '@components/ui/badge';
 import { GameTags } from '@models/CombatEditorModels';
-import { PrefixCollection } from '@utils/gameDisplayTools';
+import { PrefixCollection } from '@utils/game-display-tools';
 import { FC, ReactNode } from 'react';
 import { useTranslation } from 'react-i18next';
 import { FaTags } from 'react-icons/fa';
