+++ src/utils/editor-helpers.ts
@@ -1,12 +1,17 @@
 import { CombatEditorSaveType } from '@context/CombatEditorContext';
 import {
+    CharacterClassConversion,
     ControlledBy,
     ControlledByAI,
     ControlledByGameLogic,
     ControlledByPlayer,
     CreateCombatBody,
+    MinifiedCombatPreset,
 } from '@models/EditorConversion';
-import { isValidSquareString } from '@utils/isValidSquareString';
+
+import { CharacterDataInSave } from '@models/CombatEditorModels';
+import { CharacterInfoFull } from '@models/GameModels';
+import { isDescriptor, isValidSquareString } from '@utils/game-helpers';
 
 export const CONTROLLED_BY_PLAYER = (id: string): ControlledByPlayer => ({ type: 'player', id });
 export const CONTROLLED_BY_AI = (id: string): ControlledByAI => ({ type: 'ai', id });
@@ -99,6 +104,76 @@ class EditorHelpers {
             areaEffects: this.convertAreaEffectsToExportable(editorSave),
         };
     }
+
+    public prepareCharacterToClassConversion = (character: CharacterDataInSave): CharacterClassConversion => {
+        return {
+            decorations: character.decorations,
+            attributes: character.attributes,
+            spellBook: {
+                maxActiveSpells: character.spellBook.maxActiveSpells,
+                knownSpells: character.spellBook.knownSpells.map((spell) => ({
+                    descriptor: spell.descriptor,
+                    isActive: spell.isActive,
+                })),
+            },
+            inventory: character.inventory.map((item) => ({
+                descriptor: item.descriptor,
+                quantity: item.quantity,
+            })),
+            statusEffects: character.statusEffects.map((effect) => ({
+                descriptor: effect.descriptor,
+                duration: effect.duration,
+            })),
+            weaponry: character.weaponry.map((weapon) => ({
+                descriptor: weapon.descriptor,
+                quantity: weapon.quantity,
+            })),
+        } as CharacterClassConversion;
+    };
+
+    public minifyCharacter = (
+        character: CharacterInfoFull,
+        descriptor: string,
+    ): MinifiedCombatPreset['battlefield'][string]['character'] => {
+        return {
+            decorations: descriptor.startsWith('coordinator:')
+                ? {
+                      name: `${descriptor}.name`,
+                      description: `${descriptor}.description`,
+                      sprite: isDescriptor(character.decorations.sprite)
+                          ? character.decorations.sprite
+                          : `coordinator:${character.decorations.sprite}`,
+                  }
+                : character.decorations,
+            attributes: character.attributes,
+            spellBook: {
+                maxActiveSpells: character.spellBook.maxActiveSpells,
+                knownSpells: character.spellBook.knownSpells.map((spell) => ({
+                    descriptor: spell.descriptor,
+                    is_active: spell.isActive ?? false,
+                    turns_until_usage: spell.cooldown.current,
+                    current_consecutive_uses: spell.uses.current,
+                })),
+            },
+            inventory: character.inventory.map((item) => ({
+                descriptor: item.descriptor,
+                quantity: item.quantity,
+                turns_until_usage: item.cooldown.current,
+                current_consecutive_uses: item.uses.current,
+            })),
+            statusEffects: character.statusEffects.map((effect) => ({
+                descriptor: effect.descriptor,
+                duration: effect.duration,
+            })),
+            weaponry: character.weaponry.map((weapon) => ({
+                descriptor: weapon.descriptor,
+                quantity: weapon.quantity,
+                turns_until_usage: weapon.cooldown.current,
+                current_consecutive_uses: weapon.uses.current,
+                is_active: weapon.isActive,
+            })),
+        };
+    };
 }
 
 export default new EditorHelpers();
