+++ src/utils/game-helpers.ts
@@ -18,3 +18,19 @@ export const splitDescriptor = (full_descriptor: string): [string, string] => {
         ? (full_descriptor.split(':') as [string, string])
         : ['local', 'invalid_asset'];
 };
+export const isValidSquareString = (square: string) => {
+    return /^[1-6]\/[1-6]$/.test(square);
+};
+export const getCharacterSide = (characterLine: string | number | unknown): 'ally' | 'enemy' | null => {
+    if (typeof characterLine === 'number') {
+        return characterLine < 4 ? 'enemy' : 'ally';
+    } else if (typeof characterLine === 'string') {
+        return ['1', '2', '3'].includes(characterLine) ? 'enemy' : 'ally';
+    } else {
+        return null;
+    }
+};
+export const getCharacterSideWithSquare = (square: string): 'ally' | 'enemy' | null => {
+    const characterLine = square.split('/')[0];
+    return getCharacterSide(characterLine);
+};
