+++ src/utils/get-percentage.ts
