--- src/utils/isValidSquareString.ts
@@ -1,3 +1,0 @@
-export const isValidSquareString = (square: string) => {
-    return /^[1-6]\/[1-6]$/.test(square);
-};
